Remove MCommand, still needs a bit of work
