+++ src/main/java/com/massivecraft/factions/cmd/claim/CmdAutoClaim.java
@@ -1,6 +1,9 @@
 package com.massivecraft.factions.cmd.claim;
 
 import com.massivecraft.factions.Faction;
+import com.massivecraft.factions.cmd.CommandContext;
+import com.massivecraft.factions.cmd.CommandRequirements;
+import com.massivecraft.factions.cmd.FCommand;
 import com.massivecraft.factions.struct.Permission;
 import com.massivecraft.factions.struct.Role;
 import com.massivecraft.factions.zcore.util.TL;
