+++ src/main/java/com/massivecraft/factions/zcore/MPlugin.java
@@ -7,6 +7,7 @@ import com.massivecraft.factions.Board;
 import com.massivecraft.factions.Conf;
 import com.massivecraft.factions.FPlayers;
 import com.massivecraft.factions.Factions;
+import com.massivecraft.factions.cmd.FCommand;
 import com.massivecraft.factions.zcore.persist.SaveTask;
 import com.massivecraft.factions.zcore.util.PermUtil;
 import com.massivecraft.factions.zcore.util.Persist;
@@ -52,11 +53,12 @@ public abstract class MPlugin extends JavaPlugin {
     private MPluginSecretPlayerListener mPluginSecretPlayerListener;
 
     // Our stored base commands
-    private List<MCommand<?>> baseCommands = new ArrayList<>();
+    /*
+    private List<FCommand> baseCommands = new ArrayList<>();
 
     public List<MCommand<?>> getBaseCommands() {
         return this.baseCommands;
-    }
+    }*/
 
     // holds f stuck start times
     private Map<UUID, Long> timers = new HashMap<>();
@@ -254,6 +256,7 @@ public abstract class MPlugin extends JavaPlugin {
         return true;
     }
 
+    /*
     public boolean handleCommand(CommandSender sender, String commandString, boolean testOnly) {
         return handleCommand(sender, commandString, testOnly, false);
     }
@@ -304,7 +307,7 @@ public abstract class MPlugin extends JavaPlugin {
 
     public boolean handleCommand(CommandSender sender, String commandString) {
         return this.handleCommand(sender, commandString, false);
-    }
+    }*/
 
     // -------------------------------------------- //
     // HOOKS
